requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python-headless==4.9.0.80
face_recognition==1.3.0
numpy==1.26.4
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0